moved color class
